LinearMaxwellVlasov.jl Documentation

Core.Type — Method

The function parallel gives the same answers for a given set of inputs. Calculate the key given the inputs and the (identity) operator for storing and retrieving values

LinearMaxwellVlasov.CacheOp — Method

Identity operator for storing and retreiving values from CacheDict

LinearMaxwellVlasov.ColdSpecies — Type

ColdSpecies <: AbstractFluidSpecies

Cold plasma species. Fields:

Π Plasma frequency [rad / s]

Ω Cyclotron Frequency [rad / s]

LinearMaxwellVlasov.CoupledRelativisticSpecies — Type

Kinetic plasma species defined by one coupled distribution function in momentum space such that the relativistic dielectric tensor can be calculated. Fields:

Π Plasma frequency [rad / s]

Ω Cyclotron Frequency [rad / s]

mass Species particle mass [kg]

F :: AbstractFRelativisticMomentum Distribution function in momentum space parallel and perpendicular to the background magnetic field (normalised)

LinearMaxwellVlasov.CoupledVelocitySpecies — Type

Kinetic plasma species defined by one coupled distribution function in velocity space, parallel and perpendicular to the background magnetic field Fields:

Π Plasma frequency [rad / s]

Ω Cyclotron Frequency [rad / s]

F :: AbstractCoupledVelocity Distribution function in velocity space parallel and perpendicular to the background magnetic field (normalised)

LinearMaxwellVlasov.FParallelDiracDelta — Type

The Dirac-delta function parallel distribution functions

LinearMaxwellVlasov.FParallelNumerical — Type

Type for an arbitrary parallel distribution function that holds, a function for the distribution function, it's derivative and the limits of integral

LinearMaxwellVlasov.FPerpendicularDiracDelta — Type

The Dirac-delta function parallel distribution functions

LinearMaxwellVlasov.FPerpendicularNumerical — Type

Type for an arbitrary perpendicular distribution function that holds, a function for the distribution function, it's derivative and the limits of integral

LinearMaxwellVlasov.MaxwellianIntegralsParallel — Method

The parallel integral of the Beam only requires an integral over a drifting Maxwellian subject to the relevant kernels. All this is calculated here.

LinearMaxwellVlasov.MaxwellianIntegralsPerpendicular — Method

Perpendicular integrals for a Maxwellian

LinearMaxwellVlasov.PerpendicularKernel — Type

The kernel of the perpendicular integral

LinearMaxwellVlasov.RelativisticMaxwellian — Type

RelativisticMaxwellian Return a drifting Maxwellian function.

Members

pthz::Real - thermal momentum parallel to magnetic field [kg m/s] pth⊥::Real - thermal momentum perpendicular to magnetic field [kg m/s] pzdrift::Real=0.0 - drift parallel to the magnetic field [kg m/s] lognormalisation::Real - the log of the normalisation constant

LinearMaxwellVlasov.SeparableVelocitySpecies — Type

Kinetic plasma species with separable distribution functions parallel and perpendicular to the magnetic field. Fields:

Π Plasma frequency [rad / s]

Ω Cyclotron Frequency [rad / s]

Fz :: AbstractFParallel Distribution function parallel to magnetic field (normalised)

F⊥ :: AbstractFPerpendicular Distribution function perpendicular to magnetic field (normalised)

LinearMaxwellVlasov.ShiftedMaxwellianParallel — Type

The probability density at parallel velocity, v, of a shifted maxwellian from thermal velocity vth, and drift velocity vd.

LinearMaxwellVlasov.ShiftedMaxwellianPerpendicular — Type

The probability density at perpendicular velocity, v, of a shifted maxwellian from thermal velocity vth, and drift velocity vd.

LinearMaxwellVlasov.TransformFromInfinity — Type

Transform a function from domain [-∞, ∞]ⁿ down to [-1, 1]ⁿ

LinearMaxwellVlasov.UnitSemicircleIntegrandTransform — Type

Transform a function to transform an integrand from domain [-∞, 0]×[∞, ∞] down to [-1, -π/2]×[1, π/2].

Example:

julia> f(x) = exp(-(x[1]^2 + x[2]^2)/2) * cos(x[2])^2 * sin(x[1])^2
julia> hcubature(f, [-12.0, 0.0], [12.0, 12.0])
(0.7710130943379178, 1.1482318484139944e-8)
julia> hcubature(UnitSemicircleIntegrandTransform(f, 2.0), [0, -π/2], [1, π/2])
(0.7710130940919337, 1.1464196819507393e-8)

LinearMaxwellVlasov.WarmSpecies — Type

Warm plasma species, with speeds of sound parallel and perpendicular to the magnetic field. Fields:

Π Plasma frequency [rad / s]

Ω Cyclotron Frequency [rad / s]

soundspeed Sound speed [m/s]

LinearMaxwellVlasov.WarmSpecies — Method

WarmSpecies(Π::Float64,Ω::Float64,thermalspeed::Float64,adiabiaticindex::Number)

Warm plasma species - accept thermalspeed and ratio of specific heats to get sound speed

...

Arguments

Π: Plasma frequency [rad / s]

Ω: Cyclotron Frequency [rad / s]

thermalspeed: Thermal speed of Maxwellian distribution

adiabiaticindex: Equation of state Gruneisen gamma (ratio of specific heats)

...

LinearMaxwellVlasov.Wavenumber — Type

Wavenumber decomposed into parallel and perpendicular components Construct with parallel and wavenumber components, or by keyword arguement pairs

parallel & perpendicular

wavenumber & propagationangle

Base.Filesystem.cp — Method

Regarding integrals with besselj(n, x)* besselj(n±1, x), e.g.: where m >= 0

besselj(-m,x) * besselj(-m-1,x) == -besselj(m,x) * besselj(m+1,x)

besselj(-m+1,x) * besselj(-m-1,x) == +besselj(m-1,x) * besselj(m+1,x)

LinearMaxwellVlasov.Jμν — Method

The non-integer argument to the BesselJ

LinearMaxwellVlasov.MaxwellianSpecies — Function

MaxwellianSpecies(Π,Ω,vthb,vth⊥=vthb,vdb=0.0)

Kinetic Maxwellian Plasma species that can optionally have a drift along the magnetic field

...

Arguments

Π: Plasma frequency [rad / s]

Ω: Cyclotron Frequency [rad / s]

vthb: parallel thermal speed [m/s]

vth⊥=vthb: perpendicular thermal speed [m/s]

vdb=0.0: parallel beam speed [m/s]

Returns

SeparableVelocitySpecies(Π, Ω, FBeam(vthb, vdb), FPerpendicularMaxwellian(vth⊥))

...

Example

LinearMaxwellVlasov.RingBeamSpecies — Function

Create a kinetic plasma species with separable distribution functions parallel $f(v_\parallel)$ and perpendicular $f(v_\perpendicular)$ to the magnetic field, which are defined as a drifting beam and a ring respectively. ...

Arguments

Π: Plasma frequency [rad / s]

Ω: Cyclotron Frequency [rad / s]

vthb: parallel thermal speed [m/s]

vth⊥=vthb: perpendicular thermal speed [m/s]

vdb=0.0: parallel beam speed [m/s]

vd⊥=0.0: perpendicular ring speed [m/s]

Returns

SeparableVelocitySpecies(Π, Ω, FBeam(vthb, vdb), FRing(vth⊥, vd⊥))

...

Example

LinearMaxwellVlasov.alfvenspeed — Method

Calculate the Alfven speed given Π_Ωs, which is an iterable container of the ratio of the plasma to the cyclotron frequency of all the species of the plasma

LinearMaxwellVlasov.contribution — Function

Calculate the unitless susceptibility tensor for a cold plasma species

LinearMaxwellVlasov.contribution — Function

Calculate the converged unitless susceptibility tensor contribution for a maxwellian distribution function, having summed over bessel indices n.

LinearMaxwellVlasov.contribution — Function

Calculate the unitless susceptibility tensor for a warm plasma species Swanson 3.63

LinearMaxwellVlasov.contribution — Function

Calculate the converged unitless susceptibility tensor contribution for a maxwellian distribution function, having summed over bessel indices n.

LinearMaxwellVlasov.contribution — Method

Calculate the unitless susceptibility tensor for a maxwellian distribution function given the bessel indices n.

LinearMaxwellVlasov.dielectric — Function

The dielectric tensor for a given plasma

LinearMaxwellVlasov.dielectric — Function

The dielectric tensor for a given species

LinearMaxwellVlasov.electrostaticdielectric — Function

The electrostatic dielectric tensor for a given plasma

LinearMaxwellVlasov.foldnumeratoraboutpole — Method

Takes a function that when integrated between -Inf and +Inf returns value x, and returns a new function that returns x when integrated between real(pole) and +Inf.

LinearMaxwellVlasov.integrate — Function

Integration of Abstract Arbitrary FParallel

LinearMaxwellVlasov.integrate — Method

The integrals over the distribution functions and associated integral kernel We avoid doing integrals involving the derivative of the Dirac delta function by doing integral by parts and knowing that f-> 0 and the integral limits

LinearMaxwellVlasov.integrate — Method

Calculate the parallel integral with principal part and residue

LinearMaxwellVlasov.integrate — Method

The integrals over the distribution functions and associated integral kernel We avoid doing integrals involving the derivative of the Dirac delta function by doing integral by parts and knowing that f-> 0 and the integral limits

LinearMaxwellVlasov.integrate — Method

Integrate over the parallel distribution function multiplied by various kernels. If the imaginary part of the pole is zero, then do a trick that folds over the integral from the left of the pole to right. We need to take into account whether the real part of the pole is negative, otherwise positive slopes at negative velocities give rise to instability and not damping.

LinearMaxwellVlasov.integrate — Method

Integration of Abstract Arbitrary FPerpendicular

LinearMaxwellVlasov.limitsfolder — Method

Transform the limits of an integrand quadrature(foldnumeratoraboutpole(integrand, pole), limitsfolder(limits, pole)...)

LinearMaxwellVlasov.normalise — Method

Tool to normalize a function f between two integral limits a and b

LinearMaxwellVlasov.parallel — Method

Interface The parallel integral of the distribution function and the relevant kernel

LinearMaxwellVlasov.parallel — Method

Parallel integrals for the Beam - used for testing purposes

LinearMaxwellVlasov.parallel — Method

Compile the kernels of the parallel integral and fetch it from the DistributionFunctions module. If k parallel is zero then there is no Landau damping, so this case is made separate, as is easier to deal with.

LinearMaxwellVlasov.parallel_integral — Method

Memoise the inputs to the integrals; export the memoised function, and the dictionary of inputs -> outputs It only matters what n-m is, not the values of n and m individually So change (n, m)->(n-m, 0) to do fewer calculations

LinearMaxwellVlasov.perpendicular — Method

Interface The perpendicular integral of the distribution function and the relevant kernel

LinearMaxwellVlasov.perpendicular — Method

Calculate the definite integral kernels of the perpenedicular distribution function, or derivative thereof, multiplied by the kernel functions

LinearMaxwellVlasov.perpendicular_integral — Function

Memoise the inputs to the integrals; export the memoised function, and the dictionary of inputs -> outputs besselj harmonic numbers n and l can be any order Use this to do only half the calculatons (n, l)->(min(n, l), max(n, l)) Also know that besselj(-n, x) = (-1)^n * besselj(n, x)

LinearMaxwellVlasov.plasma_dispersion_function — Method

Return the value of the plasma dispersion function This implementation includes the residue, which is easy to verify because Z(0) = im sqrt(π).

x is the argument to the plasma disperion function

power is the moment of the integral

[1] S.D. Baalrud, Phys. Plasmas 20, 012118 (2013) and put ν = -Inf [2] M. Sampoorna et al., Generalized Voigt functions and their derivatives, Journal of Quantitative Spectroscopy & Radiative Transfer (2006), doi:10.1016/j.jqsrt.2006.08.011

LinearMaxwellVlasov.principalpart — Method

Expressing f(x) = ∑ᵢ aᵢ (x - p)ⁱ find a₋₁

LinearMaxwellVlasov.principalpartadaptive — Method

Expressing f(x) = ∑ᵢ aᵢ (x - p)ⁱ find a₋₁

LinearMaxwellVlasov.tensor — Function

Calculate the tensor representing the linear Maxwell-Vlasov set of equations. The determinant is zero when the wavenumber and frequency represent a solution to the linear Maxwell-Vlasov system of equations for these species.

LinearMaxwellVlasov.transformaboutroots — Method

Concertina and rescale a function in the sections between its roots such that all roots lie at -1 or +1. Integration over -1..1 of the resulting function gives the same answer as integration over original -Inf..Inf domain

LinearMaxwellVlasov.zerobetamagnetoacousticfrequency — Method

Eq. 25 of R. O. Dendy, C. N. Lashmore-Davies, and K. G. McClements, G. A. Cottrell, The excitation of obliquely propagating fast Alfven waves at fusion ion cyclotron harmonics, Phys. Plasmas 1 (6), June 1994

Index

Core.Type

LinearMaxwellVlasov.CacheOp

LinearMaxwellVlasov.ColdSpecies

LinearMaxwellVlasov.CoupledRelativisticSpecies

LinearMaxwellVlasov.CoupledVelocitySpecies

LinearMaxwellVlasov.FParallelDiracDelta

LinearMaxwellVlasov.FParallelNumerical

LinearMaxwellVlasov.FPerpendicularDiracDelta

LinearMaxwellVlasov.FPerpendicularNumerical

LinearMaxwellVlasov.MaxwellianIntegralsParallel

LinearMaxwellVlasov.MaxwellianIntegralsPerpendicular

LinearMaxwellVlasov.PerpendicularKernel

LinearMaxwellVlasov.RelativisticMaxwellian

LinearMaxwellVlasov.SeparableVelocitySpecies

LinearMaxwellVlasov.ShiftedMaxwellianParallel

LinearMaxwellVlasov.ShiftedMaxwellianPerpendicular

LinearMaxwellVlasov.TransformFromInfinity

LinearMaxwellVlasov.UnitSemicircleIntegrandTransform

LinearMaxwellVlasov.WarmSpecies

LinearMaxwellVlasov.WarmSpecies

LinearMaxwellVlasov.Wavenumber

Base.Filesystem.cp

LinearMaxwellVlasov.Jμν

LinearMaxwellVlasov.MaxwellianSpecies

LinearMaxwellVlasov.RingBeamSpecies

LinearMaxwellVlasov.alfvenspeed

LinearMaxwellVlasov.contribution

LinearMaxwellVlasov.contribution

LinearMaxwellVlasov.contribution

LinearMaxwellVlasov.contribution

LinearMaxwellVlasov.contribution

LinearMaxwellVlasov.dielectric

LinearMaxwellVlasov.dielectric

LinearMaxwellVlasov.electrostaticdielectric

LinearMaxwellVlasov.foldnumeratoraboutpole

LinearMaxwellVlasov.integrate

LinearMaxwellVlasov.integrate

LinearMaxwellVlasov.integrate

LinearMaxwellVlasov.integrate

LinearMaxwellVlasov.integrate

LinearMaxwellVlasov.integrate

LinearMaxwellVlasov.limitsfolder

LinearMaxwellVlasov.normalise

LinearMaxwellVlasov.parallel

LinearMaxwellVlasov.parallel

LinearMaxwellVlasov.parallel

LinearMaxwellVlasov.parallel_integral

LinearMaxwellVlasov.perpendicular

LinearMaxwellVlasov.perpendicular

LinearMaxwellVlasov.perpendicular_integral

LinearMaxwellVlasov.plasma_dispersion_function

LinearMaxwellVlasov.principalpart

LinearMaxwellVlasov.principalpartadaptive

LinearMaxwellVlasov.tensor

LinearMaxwellVlasov.transformaboutroots

LinearMaxwellVlasov.zerobetamagnetoacousticfrequency